chronos/requirements.txt
fastapi==0.111.0
typeguard==2.13.3
ddtrace==1.18.0
configparser==7.0.0
pydantic==2.7.1
pyhocon==0.3.60
aiomysql==0.2.0
starlette==0.37.2
tortoise-orm==0.21.2
boto3==1.34.125
kafka-python==2.0.2
confluent_kafka==2.2.0
requests==2.32.3
loguru==0.7.2
opentelemetry-sdk==1.38.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-httpx==0.59b0
opentelemetry-instrumentation-logging==0.59b0
cryptography==41.0.0